some post creating refactoring
